vo_opengl_cb: support display sync
The API user needs to set the "display-fps" option, or it won't work.

We need to make sure explicitly repeated frames are really accounted
for. Add a dummy mp_image struct, instead of relying that NULL values
are pushed through. May fix interpolation too, as it doesn't discard the
timing info (well, maybe). All in all, this is very dirty, and this code
really needs to be cleaned up. But for now make it work.