video: handle jittering timestamps better
For dealing with crap like Matroska. (VFR Matroska, in particular.)

This now uses an actual history of frame durations (reused from the PTS
history for the "estimated-vf-fps" property and frame backstepping). If
the durations seem to jitter out of range, stabilize_frame_duration()
signals unknown frame duration by returning -1.

Note that the --demuxer-mkv-fix-timestamps option can interfere with
correct operation of this code. It can lead to changing decisions,
which enables/disables display timing all the time.